requirements.txt
ollama
openai
pillow
python-dotenv
